// src/app/models/educationForm.ts
export interface EducationForm {
    personType: number;
    gender: number;
    nationality: string;
    stateId: string;
    mosqueId: string;
    airport: string;
    isPrice: boolean;
    studentNameSurname: string;
    studentbirthdate: string;
    studentGender: string;
    studentPassportNumber: string;
    studentIdentificationNumber:string;
    studentAddress: string;
    studentMobilePhoneNumberGermany: string;
    studentMobilePhoneNumberTurkey: string;
    transferFullName: string;
    tranferDate: string;
    transferTransactionNumber: string;
    studentFatherNameSurname: string;
    studentFatherPhone: string;
    studentMotherNameSurname: string;
    studentMotherPhone: string;
    mosqueReligiousOfficialFullName: string;
    mosqueReligiousOfficialPhone: string;  
    pasaportPath:string;
    identificationPath:string;
    studentMail:string;
    thePersonWhoRecorded:string;
  }
  